before i break down the whole app

+++ src/lib/r2.ts
@@ -15,7 +15,7 @@ const S3 = new S3Client({
 
 import crypto from 'crypto'
 import db from './db'
-import { userMedicalFilesTable } from './db/schema'
+import { userMedicalFilesTable, userTable } from './db/schema'
 import { revalidatePath } from 'next/cache'
 import { sql } from 'drizzle-orm'
 
@@ -23,15 +23,18 @@ const generateFileName = (bytes = 32) => crypto.randomBytes(bytes).toString('hex
 
 const maxFileSize = 5 * 1024 * 1024 // 5MB
 
+const acceptedPfpTypes = ['image/jpg', 'image/png', 'image/jpeg', 'image/webp']
 const acceptedTypes = ['image/jpg', 'image/png', 'image/jpeg', 'image/webp', 'application/pdf']
 
-export async function getSignUrl(key: string | '', type: string, size: number, checkSum: string, folderId?: number) {
+export async function getSignUrl(key: string | '', type: string, size: number, checkSum?: string, folderId?: number, userId?: string, lastPic?: string) {
     const { user } = await validateRequest()
 
     if (!user) throw new Error('Unauthorized')
 
     if (size > maxFileSize) throw new Error('File size too large')
-    if (!acceptedTypes.includes(type)) throw new Error('Unsupported file type')
+
+    if (userId && !acceptedPfpTypes.includes(type)) throw new Error('Unsupported file type')
+    if (!userId && !acceptedTypes.includes(type)) throw new Error('Unsupported file type')
 
     const fileName = generateFileName() + `-${key}`;
     const putObjectCommand = new PutObjectCommand({
@@ -39,7 +42,7 @@ export async function getSignUrl(key: string | '', type: string, size: number, c
         Key: fileName,
         ContentType: type,
         ContentLength: size,
-        ChecksumSHA256: checkSum,
+        ChecksumSHA256: checkSum ? checkSum : undefined,
         Metadata: {
             // so we can check which user uploaded it later in the frontend
             userId: user.id,
@@ -49,19 +52,25 @@ export async function getSignUrl(key: string | '', type: string, size: number, c
     const signedUrl = await getSignedUrl(S3, putObjectCommand, { expiresIn: 3600 })
 
     if (folderId) {
+        console.log('folderId', folderId);
         await db.insert(userMedicalFilesTable).values({ name: fileName, folderId: folderId });
+    }
+
+    if (userId) {
+        console.log('userId', userId)
+        await db.update(userTable).set({ picture: fileName }).where(sql`${userTable.id} = ${userId}`);
     }
 
     return {
         success: {
             url: signedUrl,
             fileName,
-            folderId: folderId
         }
     }
 }
 
-export async function deleteFiles(names: { name: string, folderId: number }[], s3: boolean = false) {
+export async function deleteFiles(names: string[], s3: boolean = false) {
+    console.log(names)
     const { user } = await validateRequest()
     if (!user) throw new Error('Unauthorized')
 
@@ -82,41 +91,42 @@ export async function deleteFiles(names: { name: string, folderId: number }[], s
 
 
     for (const fileName of names) {
-        const { name } = fileName;
-
         // Find the file in the user's folders
         const fileFound = userFolders.folders.some(folder =>
-            folder.files.some(file => file.name === name)
+            folder.files.some(file => file.name === fileName)
         );
 
         if (fileFound) {
+            console.log('file found')
             if (s3) {
-                console.log('lmfao work plz')
+                console.log('s3 found')
                 const deleteObjectCommand = new DeleteObjectCommand({
                     Bucket: process.env.R2_BUCKET_NAME || '',
-                    Key: name
+                    Key: fileName
                 });
 
                 try {
+                    console.log('deleteObjectCommand')
                     await S3.send(deleteObjectCommand);
                 } catch (error) {
-                    console.error(`Error deleting file from S3: ${name}`, error);
+                    console.error(`Error deleting file from S3: ${fileName}`, error);
                     // Handle S3 deletion error (e.g., logging, retrying, etc.)
                 }
             }
 
             try {
-                await db.delete(userMedicalFilesTable).where(sql`${userMedicalFilesTable.name} = ${name}`);
+                console.log('try delete from db')
+                await db.delete(userMedicalFilesTable).where(sql`${userMedicalFilesTable.name} = ${fileName}`);
             } catch (error) {
-                console.error(`Error deleting file from database: ${name}`, error);
+                console.error(`Error deleting file from database: ${fileName}`, error);
             }
         } else {
-            console.warn(`File not found in user's folders: ${name}`);
+            console.warn(`File not found in user's folders: ${fileName}`);
         }
     }
 }
 
-export async function deleteFile(name: string, s3: boolean = false) {
+export async function deleteFile(name: string, file: boolean = true, s3: boolean = false) {
     const { user } = await validateRequest()
     if (!user) throw new Error('Unauthorized')
 
@@ -135,8 +145,15 @@ export async function deleteFile(name: string, s3: boolean = false) {
     }
 
     try {
-        await db.delete(userMedicalFilesTable).where(sql`${userMedicalFilesTable.name} = ${name}`);
+        if (file) {
+            await db.delete(userMedicalFilesTable).where(sql`${userMedicalFilesTable.name} = ${name}`);
+        }
     } catch (error) {
         console.error(`Error deleting file from database: ${name}`, error);
     }
+}
+
+export async function uploaded(path: string) {
+    if (!path) throw new Error('Invalid path')
+    return revalidatePath(path)
 }
